Ajout de la condition gérant le choix entre playlist et chaîne
