--- scripts/duration/duration_playlist.py
@@ -1,23 +1,0 @@
-from pytube import Playlist, Channel
-
-url = input("Donne moi le lien de la playlist ! ") # Demande le lien de la playlist
-playlist = Playlist(url) # Crée l'objet playlist à partir du lien donné
-
-# Boucle sur les vidéos de la playlist pour calculer la durée totale de la playlist
-duree_totale = 0
-for video in playlist.videos:
-    duree_totale += video.length
-
-# Fonction permettant de calculer la durée
-def duree(duree_totale_seconde):
-    heure = duree_totale_seconde // 3600
-    duree_totale_seconde -= (duree_totale_seconde // 3600) * 3600 # Retire les secondes correspondant au nombre d'heures de la durée totale
-    minute = duree_totale_seconde // 60
-    duree_totale_seconde -= (duree_totale_seconde // 60) * 60 # Retire les secondes correspondant au nombre de minutes de la durée totale
-    seconde = duree_totale_seconde # La durée totale restante correspond aux secondes (variable crée pour une meilleure lisibilitée)
-    return heure, minute, seconde
-
-duree_playlist = duree(duree_totale) # Calcule la durée de la playlist
-duree_moy_video = duree(round(duree_totale / len(playlist.videos))) # Calcule la durée moyenne d'une vidéo de la playlist
-print(f"La durée de la playlist est de {duree_playlist[0]}h{duree_playlist[1]}m{duree_playlist[2]}s.") # Informe de la durée totale de la playlist
-print(f"Sur les {len(playlist.videos)} vidéos de la playlist, une vidéo dure en moyenne :\n{duree_moy_video[0]}h{duree_moy_video[1]}m{duree_moy_video[2]}s")
